requirements.txt
asgiref==3.5.0
Django==4.0.3
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
phonenumberslite==8.12.45
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
gunicorn==20.1.0
psycopg2-binary==2.8.5